api/pyproject.toml
[tool.poetry]
name = "ai-agent-template"
version = "0.1.0"
description = ""
authors = ["Sumit Rathore <sumit@neolen.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.32.3"
boto3 = "^1.34.145"
gradio = "^4.38.1"
langchain = "^0.2.10"
langchain-openai = "^0.1.17"
langchain-aws = "^0.1.11"
langgraph = "^0.1.14"
weaviate-client = "^4.7.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
